+++ airflow/gcp/utils/__init__.py
@@ -16,19 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.gcp.operators.mlengine`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.gcp.operators.mlengine import (  # noqa
-    MLEngineBatchPredictionOperator,
-    MLEngineModelOperator,
-    MLEngineTrainingOperator,
-    MLEngineVersionOperator,
-)
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.gcp.operators.mlengine`.",
-    DeprecationWarning,
-)
