+++ scripts/ci/pylint_todo.txt
@@ -66,7 +66,7 @@
 ./airflow/contrib/operators/jenkins_job_trigger_operator.py
 ./airflow/contrib/operators/jira_operator.py
 ./airflow/contrib/operators/kubernetes_pod_operator.py
-./airflow/contrib/operators/mlengine_operator.py
+./airflow/gcp/operators/mlengine.py
 ./airflow/contrib/operators/mongo_to_s3.py
 ./airflow/contrib/operators/opsgenie_alert_operator.py
 ./airflow/contrib/operators/oracle_to_azure_data_lake_transfer.py
@@ -135,8 +135,8 @@
 ./airflow/contrib/sensors/wasb_sensor.py
 ./airflow/contrib/sensors/weekday_sensor.py
 ./airflow/contrib/utils/log/task_handler_with_custom_formatter.py
-./airflow/contrib/utils/mlengine_operator_utils.py
-./airflow/contrib/utils/mlengine_prediction_summary.py
+./airflow/gcp/utils/mlengine_operator_utils.py
+./airflow/gcp/utils/mlengine_prediction_summary.py
 ./airflow/contrib/utils/sendgrid.py
 ./airflow/contrib/utils/weekday.py
 ./airflow/exceptions.py
@@ -413,7 +413,7 @@
 ./tests/contrib/operators/test_emr_create_job_flow_operator.py
 ./tests/contrib/operators/test_gcs_to_s3_operator.py
 ./tests/contrib/operators/test_grpc_operator.py
-./tests/contrib/operators/test_mlengine_operator.py
+./tests/gcp/operators/test_mlengine.py
 ./tests/contrib/operators/test_oracle_to_azure_data_lake_transfer.py
 ./tests/contrib/operators/test_qubole_check_operator.py
 ./tests/contrib/operators/test_redis_publish_operator.py
@@ -435,7 +435,7 @@
 ./tests/contrib/sensors/test_weekday_sensor.py
 ./tests/contrib/utils/logging_command_executor.py
 ./tests/contrib/utils/run_once_decorator.py
-./tests/contrib/utils/test_mlengine_operator_utils.py
+./tests/gcp/operators/test_mlengine_utils.py
 ./tests/contrib/utils/test_sendgrid.py
 ./tests/contrib/utils/test_weekday.py
 ./tests/core.py
